refactor(#17): move tag editor component to shared dir
